requirements.txt
flask
flask-cors
openai
google-generativeai
python-docx
python-dotenv
qrcode[pil]
pillow
gunicorn
